Bugfixes for hotel occupancy calculation
The calculation can't be done on a post_save signal as there is no way
of knowing from wich hotel the request was removed, if any.
Also the new calculation is less resource intensive, as it needs the
changed request and max two hotels from the db, whereas the old one had
to filter over all requests.